put bsort into a header file and reference it from bsort.c
